chore: optimize local dev of tui
