+++ package.json
@@ -14,9 +14,7 @@
     "build:test": "TEST=1 vp run build",
     "dev": "nuxt dev",
     "dev:docs": "vp run --filter npmx-docs dev --port=3001",
-    "npmx-tui": "vp run --filter npmx-tui dev:local",
-    "npmx-tui:solo": "vp run --filter npmx-tui dev:ffi",
-    "npmx-tui:watch": "vp run --filter npmx-tui dev:watch",
+    "npmx-tui": "vp run --filter npmx-tui dev",
     "i18n:check:fix": "node scripts/compare-translations.ts --fix",
     "i18n:report:fix": "node scripts/remove-unused-translations.ts",
     "knip:fix": "knip --fix",
