+++ tui/package.json
@@ -24,11 +24,8 @@
   },
   "scripts": {
     "build": "vp pack",
-    "dev": "node scripts/dev-local.ts",
-    "dev:local": "node scripts/dev-local.ts",
-    "dev:ffi": "node --experimental-ffi src/cli.ts",
-    "dev:once": "node src/cli.ts",
-    "dev:watch": "node scripts/dev.ts",
+    "dev": "node src/cli.ts",
+    "dev:watch": "node --watch src/cli.ts",
     "test:types": "tsc --noEmit"
   },
   "dependencies": {
